backend/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
transformers
torch
nltk
youtube-transcript-api
google-api-python-client
python-dotenv
fastapi
uvicorn
